requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-environ==0.9.0
django-formtools==2.4.1
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.3.0
